Move theme selectors
